create new file for LinkedList class
